NoiseModel moved to its own module
